python/pyproject.toml
[project]
name = "mirascope"
version = "2.0.0-alpha.3"
description = "LLM abstractions that aren't obstructions"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "William Bakst", email = "william@mirascope.com" },
    { name = "Dandelion Mané", email = "dandelion@mirascope.com" },
]
maintainers = [
    { name = "William Bakst", email = "william@mirascope.com" },
    { name = "Dandelion Mané", email = "dandelion@mirascope.com" },
]
keywords = [
    "agents",
    "artificial intelligence",
    "developer tools",
    "llm",
    "llm tools",
    "prompt engineering",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: File Formats :: JSON",
    "Topic :: File Formats :: JSON :: JSON Schema",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "docstring-parser>=0.17.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.10.0",
]

[project.urls]
Homepage = "https://mirascope.com"
Documentation = "https://mirascope.com/docs/mirascope/v2"
Repository = "https://github.com/Mirascope/mirascope/tree/v2"
Issues = "https://github.com/Mirascope/mirascope/issues"
Changelog = "https://github.com/Mirascope/mirascope/releases"

[project.optional-dependencies]
anthropic = ["anthropic>=0.72.0,<1.0"]
api = ["pydantic-settings>=2.12.0"]
google = ["google-genai>=1.48.0,<2", "pillow>=10.4.0,<12", "proto-plus>=1.24.0"]
openai = ["openai>=2.7.1,<3"]
mcp = ["mcp>=1.0.0,<2"]
ops = [
    "opentelemetry-sdk>=1.38.0,<2",
    "opentelemetry-api>=1.38.0,<2",
    "opentelemetry-instrumentation>=0.59b0,<1",
    "opentelemetry-exporter-otlp>=1.38.0,<2",
    "opentelemetry-propagator-b3>=1.38.0,<2",
    "orjson>=3.11.4",
    "opentelemetry-propagator-b3>=1.38.0",
    "opentelemetry-propagator-jaeger>=1.38.0",
    "libcst>=1.8.6",
    "packaging>=25.0",
]
mlx = ["mlx-lm>=0.28.4,<1"]
together = ["together>=1.5.31,<2"]
all = ["mirascope[anthropic,api,google,openai,mcp,ops,mlx,together]"]

[dependency-groups]
dev = [
    "ruff>=0.9.7",
    "pytest>=8.3.2",
    "pyright>=1.1.407",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.8.0",
    "codespell>=2.3.0",
    "vcrpy>=7.0.0",
    "inline-snapshot>=0.26.0",
    "pytest-vcr>=1.0.2",
    "pytest-httpserver>=1.1.0",
    "datamodel-code-generator>=0.39.0",
]


[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.inline-snapshot]
format-command = "ruff check --fix-only --stdin-filename {filename} | ruff format --stdin-filename {filename}"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "venv",
    ".coveragerc",
]
line-length = 88
target-version = "py310"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"examples/*.{py}" = ["T201", "ANN"]
"scripts/*.{py}" = ["T201"]
"typechecking/*.{py}" = ["T201", "ANN"]
"mirascope/api/_generated/**" = ["ALL"]
"tests/ops/_internal/closure_test_functions/**" = ["ANN"]
"tests/ops/_internal/test_closure.py" = ["ANN"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "YTT",
    "B",
    "T10",
    "T20",
    "C4",
    "PYI006",
    "SIM",
    "ANN",
    "RUF022",
]
ignore = ['D105', 'D107', 'D205', 'D415', 'E501', 'B011', 'B028', 'B904']
fixable = ["ALL"]
pydocstyle = { convention = 'google' }
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ['mirascope', 'tests', 'examples']
extra-standard-library = ["typing_extensions"]
combine-as-imports = true

[tool.pyright]
include = ["mirascope", "tests", "examples"]
exclude = [".venv", "mirascope/api/_generated", "mirascope/llm/mcp"]
venvPath = "."
venv = ".venv"
stubPath = "typings"
reportMatchNotExhaustive = "error"
reportMissingParameterType = "error"
reportMissingTypeArgument = "error"
reportPrivateUsage = "error"
reportUnknownArgumentType = "error"
reportUnknownMemberType = "error"
reportUnknownParameterType = "error"
reportUnknownVariableType = "error"
reportUnnecessaryCast = "error"
reportUnnecessaryComparison = "error"
reportUnnecessaryTypeIgnoreComment = "error"
reportUntypedClassDecorator = "error"
reportUntypedFunctionDecorator = "error"
reportUnusedClass = "error"
reportUnusedFunction = "error"
reportUnusedImport = "error"
reportUnusedVariable = "error"
strictListInference = true

[[tool.pyright.executionEnvironments]]
root = "examples"
reportMatchNotExhaustive = "none"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"

[[tool.pyright.executionEnvironments]]
root = "tests"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"

[[tool.pyright.executionEnvironments]]
root = "typechecking"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"
reportUnusedVariable = "none"

[[tool.pyright.executionEnvironments]]
root = "mirascope/ops"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnusedImport = "none"
reportPrivateUsage = "none"

[tool.datamodel-codegen]
input-file-type = "jsonschema"
output-model-type = "typing.TypedDict"
use-standard-collections = true
use-union-operator = true
use-inline-field-description = true
use-type-alias = true
disable-timestamp = true
keep-model-order = true
reuse-model = true
reuse-scope = "tree"
all-exports-scope = "recursive"
all-exports-collision-strategy = "error"
custom-file-header = "\"\"\"OpenTelemetry Gen AI Semantic Conventions types.\"\"\""
formatters = ["ruff-check", "ruff-format"]
target-python-version = "3.10"
type-mappings = ["binary=string"]

[tool.codespell]
skip = [".git", "*.lock"]

[tool.coverage.run]
omit = []

[tool.coverage.report]
exclude_also = ["@overload", "@abstractmethod", "if TYPE_CHECKING:"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
